Add specific class for custom defaults
